python/requirements.txt
# FlightSync - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0

# Database Drivers
psycopg2-binary>=2.9.9
pymongo>=4.6.0

# Data Validation
pydantic>=2.5.0
pydantic[email]>=2.5.0

# Environment Configuration
python-dotenv>=1.0.0

# Authentication
PyJWT>=2.8.0

# Scientific Computing (for pricing engine)
numpy>=1.26.0
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.2

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0

# Optional: For production deployment
gunicorn>=21.2.0
